pyproject.toml
[project]
authors = [
  {name = "Patrick Arminio", email = "patrick.arminio@gmail.com"},
]
dependencies = [
  "strawberry-graphql",
  "fastapi",
  "uvicorn",
  "psycopg[pool]>=3.1.9",
  "fastapi-sessions>=0.3.2",
]
description = ""
license = {text = "MIT"}
name = ""
requires-python = ">=3.11"
version = ""

[tool.black]
line-length = 88

[tool.isort]
combine_as_imports = true
combine_star = true
indent = 4
known_pytest = ["pytest"]
known_strawberry = ["strawberry"]
lines_after_imports = 2
profile = "black"
sections = ["FUTURE", "STDLIB", "PYTEST", "THIRDPARTY", "STRAWBERRY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
plugins = ["strawberry.ext.mypy_plugin"]

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = ["black", "flake8", "mypy", "pytest", "httpx"]

[tool.pdm.scripts]
dev.cmd = "uvicorn main:app --reload"
dev.env_file = ".env"
start = "uvicorn main:app --host 0.0.0.0 --port 8080"

[tool.ruff]
select = ["I", "E", "F"]
fix = true

[build-system]
build-backend = "pdm.backend"
requires = ["pdm-backend"]
